fix(refactor): Fixed errors regarding the new factory pattern
