+++ src/main/java/views/content/TileView.java
@@ -1,4 +1,4 @@
-package views;
+package views.content;
 
 
 import javafx.scene.paint.Color;
